+++ Gifting.Public/ViewModels/Common/DropdownItemViewModel.cs
@@ -2,13 +2,13 @@ namespace Gifting.Public.ViewModels.Common
 {
 	public struct DropdownItemViewModel
 	{
-		public DropdownItemViewModel(int id, string name)
+		public DropdownItemViewModel(long id, string name)
 		{
 			Id = id;
 			Name = name;
 		}
 
-		public int Id { get; set; }
+		public long Id { get; set; }
 
 		public string Name { get; set; }
 	}
